ops: moves tests from mocha to jest, adds redis for cache
