+++ docker-compose.dev.yaml
@@ -1,0 +1,15 @@
+version: '3'
+services:
+  redis:
+    container_name: redis
+    image: redis/redis-stack:6.2.6-v8
+    ports:
+      - 8888:8001
+    volumes:
+      - redis_data:/data
+
+volumes:
+  redis_data:
+
+networks:
+  line:
