+++ docker-compose.yaml
@@ -11,3 +11,19 @@ services:
       - .env
     environment:
       NODE_ENV: production
+    networks:
+      - line
+
+  redis:
+    container_name: redis
+    image: redis/redis-stack-server:6.2.6-v8
+    volumes:
+      - redis_data:/data
+    networks:
+      - line
+
+volumes:
+  redis_data:
+
+networks:
+  line:
