+++ src/libs/redis.ts
@@ -1,0 +1,5 @@
+import redis from 'redis'
+
+const client = redis.createClient()
+
+export default client
