+++ src/types/libs.ts
@@ -2,8 +2,8 @@ import { SupabaseClient } from '@supabase/supabase-js'
 import { Collection } from 'discord.js'
 import { BotConfig } from '../controllers/bot/config.controller'
 import { GuildSettings } from '../controllers/bot/guilds.controller'
-import { GuildPluginData } from '../controllers/bot/plugins.controller'
 import { Database } from './database.types'
+import { GuildPluginData } from './plugins'
 
 declare module 'discord.js' {
   export interface Client {
