+++ src/utils/dates.ts
@@ -2,12 +2,26 @@ import { differenceInSeconds, endOfDay } from 'date-fns'
 import { formatInTimeZone } from 'date-fns-tz'
 import formatDistanceToNow from 'date-fns/formatDistanceToNow'
 
+/**
+ * Returns a string representing the time difference between the given time and now.
+ * @param time - The time to format.
+ * @returns A string representing the time difference between the given time and now.
+ */
 export const formatFromNow = (time: number | Date) => formatDistanceToNow(time, { addSuffix: true })
 
+/**
+ * Returns a promise that resolves after the given number of milliseconds.
+ * @param ms - The number of milliseconds to sleep.
+ * @returns A promise that resolves after the given number of milliseconds.
+ */
 export const sleep = (ms: number) => {
   return new Promise((resolve) => setTimeout(resolve, ms))
 }
 
+/**
+ * Returns a string representing the time remaining until midnight.
+ * @returns A string representing the time remaining until midnight.
+ */
 export const getTimeRemainingUntilMidnight = () => {
   const now = new Date()
   const midnight = endOfDay(now)
@@ -24,6 +38,12 @@ export const getTimeRemainingUntilMidnight = () => {
   }
 }
 
+/**
+ * Returns the local time in two different time zones.
+ * @param targetTimezone - The target timezone to get the local time for.
+ * @param authorTimezone - The author's timezone to get the local time for.
+ * @returns An object containing the author's local time and the target timezone's local time.
+ */
 export const getTimeZonesTime = (
   targetTimezone: string,
   authorTimezone: string,
@@ -36,6 +56,11 @@ export const getTimeZonesTime = (
   return { targetLocalTime, authorLocalTime: localTime }
 }
 
+/**
+ * Extracts the hours and minutes from a given time string.
+ * @param time - The time string to extract the hours and minutes from.
+ * @returns A string representing the hours and minutes in the format "HH:mm".
+ */
 export const extractHours = (time: string) => {
   return new Date(time).toLocaleTimeString('en-US', {
     hour: '2-digit',
